+++ src/gen-charts.ts
@@ -1703,11 +1703,9 @@ function makeValAxis(opts: IChartOptsLib, valAxisId: string): string {
 			: 'between') +
 		'"/>'
 	if (opts.valAxisMajorUnit) strXml += ' <c:majorUnit val="' + opts.valAxisMajorUnit + '"/>'
-	if (opts.valAxisDisplayUnit) {
-		strXml += `<c:dispUnits><c:builtInUnit val="${opts.valAxisDisplayUnit}"/>`
-		if (opts.valAxisDisplayUnitLabel) strXml += `<c:dispUnitsLbl/>`
-		strXml += `</c:dispUnits>`
-	}
+	if (opts.valAxisDisplayUnit)
+		strXml += `<c:dispUnits><c:builtInUnit val="${opts.valAxisDisplayUnit}"/>${opts.valAxisDisplayUnitLabel ? '<c:dispUnitsLbl/>' : ''}</c:dispUnits>`
+
 	strXml += '</c:valAx>'
 
 	return strXml