+++ types/index.d.ts
@@ -1434,7 +1434,7 @@ declare namespace PptxGenJS {
 		valAxes?: number[]
 		valAxisCrossesAt?: string | number
 		valAxisDisplayUnit?: 'billions' | 'hundredMillions' | 'hundreds' | 'hundredThousands' | 'millions' | 'tenMillions' | 'tenThousands' | 'thousands' | 'trillions'
-		valAxisDisplayUnitLabel?:boolean
+		valAxisDisplayUnitLabel?: boolean
 		valAxisHidden?: boolean
 		valAxisLabelColor?: string
 		valAxisLabelFontBold?: boolean